scenarios/handson_microsoft_search/pyproject.toml
[tool.poetry]
name = "handson-microsoft-search"
version = "0.0.1"
description = ""
authors = ["ks6088ts <ks6088ts@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.9.0"
python-dotenv = "^1.0.1"
msgraph-sdk = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
